+++ content/lib/pyswitch/controller/actions/AnalogAction.py
@@ -1,24 +1,28 @@
-from ...misc import PeriodCounter
+from ...misc import PeriodCounter, Updateable
 
 from adafruit_midi.system_exclusive import SystemExclusive
 
-class AnalogAction:
+class AnalogAction(Updateable):
     
     # Use this action for all analog inputs like expression pedals.
     def __init__(self, 
-                 mapping,                          # Parameter mapping to be controlled
-                 max_frame_rate = 24,              # Maximum frame rate for sending MIDI values (fps)
-                 max_value = None,                 # Maximum value of the mapping (16383 for NRPN, 127 for CC) Set this to None to auto-detect these ranges.
-                 num_steps = 128,                  # Number of steps to be regarded as different (saves MIDI traffic at the cost of precision).
-                 enable_callback = None,           # Callback to set enabled state (optional). Must contain an enabled(action) function.
+                 mapping,                           # Parameter mapping to be controlled
+                 max_frame_rate = 24,               # Maximum frame rate for sending MIDI values (fps)
+                 max_value = None,                  # Maximum value of the mapping (16383 for NRPN, 127 for CC) Set this to None to auto-detect these ranges.
+                 num_steps = 128,                   # Number of steps to be regarded as different (saves MIDI traffic at the cost of precision).
+                 enable_callback = None,            # Callback to set enabled state (optional). Must contain an enabled(action) function.
                  id = None,
-                 auto_calibrate = True,            # Auto-calibrate (similar to Kemper devices)
-                 cal_min_window = 0.25,            # Minimum calibration range in percent [0..1]. A value of 0.25 means that the pedal has to cover at least 25% of the full range to be regarded.
-                 transfer_function = None          # Transfer function. Input is the raw encoder value in range [0..65535]. Must return the value to be sent for 
-                                                   # the mapping, in its range. If used, this disables the num_steps and max_value parameters.
-                                                   # 
-                                                   # NOTE: When designing your transfer function, take care that the output values do not change too often as this 
-                                                   # imposes performance issues! 
+                 auto_calibrate = True,             # Auto-calibrate (similar to Kemper devices)
+                 cal_min_window = 0.25,             # Minimum calibration range in percent [0..1]. A value of 0.25 means that the pedal has to cover at least 25% of the full range to be regarded.
+                 transfer_function = None,          # Transfer function. Input is the raw encoder value in range [0..65535]. Must return the value to be sent for 
+                                                    # the mapping, in its range. If used, this disables the num_steps and max_value parameters.
+                                                    # 
+                                                    # NOTE: When designing your transfer function, take care that the output values do not change too often as this 
+                                                    # imposes performance issues!
+                 preview_display = None,            # If assigned, the adjusted value will be displayed in the passed DisplayLabel when the pedal is adjusted. 
+                 preview_timeout_millis = 1500,     # This is the amount of time (milliseconds) after which the 
+                                                    # preview display will return to its normal state.
+                 convert_value = None,              # Optional conversion routine for displaying values: (value) => string
         ):
         if isinstance(mapping.set, SystemExclusive):
             if max_value == None:
@@ -40,15 +44,30 @@ class AnalogAction:
         self.__step_width = int(65536 / num_steps)
         self.__last_value = -1
         self.__transfer_function = transfer_function
+        self.__convert_value = convert_value
         
         self.__calibrate = auto_calibrate
         self.__cal_min = None
         self.__cal_max = None
-        self.__cal_min_window = int(65536 * cal_min_window)  
+        self.__cal_min_window = int(65536 * cal_min_window)
+
+        if preview_display:
+            from ..preview import ValuePreview
+
+            self.__preview = ValuePreview.get(
+                label = preview_display,
+                timeout_millis = preview_timeout_millis
+            )                
+        else:
+            self.__preview = None
 
     @property
     def enabled(self):
         return self.__enable_callback.enabled(self) if self.__enable_callback else True
+
+    def update(self):
+        if self.__preview:
+            self.__preview.update()
 
     # is_pot tells whether the input is a potentiometer (True) or a rotary encoder (False)
     def init(self, appl):
@@ -101,3 +120,14 @@ class AnalogAction:
                 self.__last_value = v
                 self.__appl.client.set(self.__mapping, v)
 
+                if self.__preview:
+                    if not self.__convert_value:
+                        self.__preview.preview_mapping(
+                            mapping = self.__mapping,
+                            value = v,
+                            max_value = self.__max_value
+                        )
+                    else:
+                        self.__preview.preview(
+                            text = self.__convert_value(v)
+                        )