+++ test/pyswitch/test_callback_binary_parameter.py
@@ -28,7 +28,7 @@ with patch.dict(sys.modules, {
     from lib.pyswitch.controller.actions import PushButtonAction
     from lib.pyswitch.controller.callbacks import BinaryParameterCallback
     
-    from lib.pyswitch.controller import Controller
+    from lib.pyswitch.controller.controller import Controller
 
     from .tools import *