+++ test/pyswitch/test_client_mapping.py
@@ -26,13 +26,11 @@ with patch.dict(sys.modules, {
 
 class TestClientParameterMapping(unittest.TestCase):
 
-    def test_singleton(self):
-        # Test the Mock Parameter Mapping unique name checker routine
+    def test_private_init(self):
         with self.assertRaises(Exception):
             ClientParameterMapping(name = "footest", create_key = 1)
 
     def test_singleton_name_check(self):
-        # Test the Mock Parameter Mapping unique name checker routine
         ClientParameterMapping.get(name = "footest")
 
         with self.assertRaises(Exception):